spring security the first commit

+++ src/main/java/com/example/normal/entity/SysUser.java
@@ -24,7 +24,7 @@ public class SysUser implements Serializable {
      * 部门ID
      */
     @TableField(value = "dept_id")
-    private Long deptId;
+    private Integer deptId;
 
     /**
      * 用户账号
